Add files via upload
Purpose: This lab will allow the student to get a basic understanding of file content manipulation using the sed and grep commands, using file permissions, and cron jobs in the Unix file system as used by a system administrator. As a system administrator you will need to be resourceful, so feel free to Google the command, use any reference text, and/or collaborate with classmates on executing the commands and seeing what results are expected and also what options can be used with each command.